backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=3.1.7,<4.0.0
python-multipart==0.0.9
face_recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pandas==2.2.1
openpyxl==3.1.2
aiofiles==23.2.1
mediapipe==0.10.9
